backend/requirements.txt
Jinja2==3.1.4
MarkupSafe==3.0.1
PyJWT==2.9.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi>=0.68.0,<0.69.0
gunicorn>=23.0.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
numpy==2.0.2
oauthlib==3.2.2
openai>=1.2.0
packaging==24.1
passlib[bcrypt]>=1.7.4,<2.0.0
pgeocode>=0.3.0
pgvector==0.3.6
platformdirs==4.3.6
psycopg2-binary>=2.9.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv>=0.19.0,<0.20.0
python-jose[cryptography]>=3.3.0,<4.0.0
python-multipart>=0.0.5,<0.1.0
redis==5.2.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlalchemy>=1.4.23,<2.0.0
supabase
tuspy==1.1.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn[standard]>=0.15.0,<0.16.0
websockets==13.1
zipp==3.20.2
